[fix] user entity 수정 및 mock data 비율 수정
User entity 프로퍼티 추가
  - email
  - role
Mock data 비율 수정
  - user 1000
  - community 300
  - communityUser 9800
  - keyword 12500
  - keywordUser 150000
  - thread 300000
